Refactor Styles
